Feat: infiniteScroll기능 재사용을 위해 useInfiniteScroll.js로 분리:
